[refactor] clusterer 객체 분리
