+++ app/src/main/java/com/squirtles/musicroad/map/NaverMap.kt
@@ -5,7 +5,6 @@ import android.app.Activity
 import android.content.Context
 import android.graphics.PointF
 import android.location.Location
-import android.view.View
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
@@ -34,29 +33,16 @@ import com.naver.maps.map.LocationTrackingMode
 import com.naver.maps.map.MapView
 import com.naver.maps.map.NaverMap
 import com.naver.maps.map.UiSettings
-import com.naver.maps.map.clustering.ClusterMarkerInfo
 import com.naver.maps.map.clustering.Clusterer
-import com.naver.maps.map.clustering.DefaultClusterMarkerUpdater
-import com.naver.maps.map.clustering.DefaultLeafMarkerUpdater
-import com.naver.maps.map.clustering.DefaultMarkerManager
-import com.naver.maps.map.clustering.LeafMarkerInfo
-import com.naver.maps.map.overlay.Align
 import com.naver.maps.map.overlay.CircleOverlay
 import com.naver.maps.map.overlay.LocationOverlay
-import com.naver.maps.map.overlay.Marker
-import com.naver.maps.map.overlay.Overlay
 import com.naver.maps.map.overlay.OverlayImage
 import com.naver.maps.map.util.FusedLocationSource
 import com.squirtles.musicroad.R
-import com.squirtles.musicroad.map.marker.ClusterMarkerIconView
-import com.squirtles.musicroad.map.marker.DensityType
-import com.squirtles.musicroad.map.marker.LeafMarkerIconView
 import com.squirtles.musicroad.map.marker.MarkerKey
-import com.squirtles.musicroad.ui.theme.Black
-import com.squirtles.musicroad.ui.theme.Blue
+import com.squirtles.musicroad.map.marker.buildClusterer
 import com.squirtles.musicroad.ui.theme.Primary
 import com.squirtles.musicroad.ui.theme.Purple15
-import com.squirtles.musicroad.ui.theme.White
 import kotlinx.coroutines.launch
 import kotlin.coroutines.resume
 import kotlin.coroutines.suspendCoroutine
@@ -88,107 +74,8 @@ fun NaverMap(
     }
 
     DisposableEffect(Unit) {
-        clusterer = Clusterer.ComplexBuilder<MarkerKey>()
-            .thresholdStrategy { zoom ->
-                when {
-                    zoom >= 16 -> 10.0
-                    zoom >= 13 -> 20.0
-                    zoom >= 11 -> 40.0
-                    else -> 70.0
-                }
-            }
-            .tagMergeStrategy { cluster ->
-                cluster.children.map { it.tag }.joinToString(",")
-            }
-            .markerManager(object : DefaultMarkerManager() {
-                override fun createMarker(): Marker {
-                    val marker = Marker()
-                    with(marker) {
-                        icon = OverlayImage.fromView(View(context))
-                        setCaptionAligns(Align.Center)
-                        captionHaloColor = android.graphics.Color.TRANSPARENT
-                    }
-                    return marker
-                }
-            })
-            .clusterMarkerUpdater(object : DefaultClusterMarkerUpdater() {
-                override fun updateClusterMarker(info: ClusterMarkerInfo, marker: Marker) {
-                    // 클릭된 마커가 클러스터 마커 안에 포함되면 클릭된 마커 해제
-                    mapViewModel.clickedMarkerState.value.curPickId?.let { curPickId ->
-                        if (info.tag.toString().contains(curPickId)) {
-                            mapViewModel.resetClickedMarkerState(context)
-                        }
-                    }
-                    val densityType = when {
-                        info.size < 10 -> DensityType.LOW
-                        info.size < 100 -> DensityType.MEDIUM
-                        else -> DensityType.HIGH
-                    }
-                    val captionColor = when {
-                        densityType == DensityType.LOW -> Black
-                        else -> White
-                    }
-                    val clusterMarkerIconView = ClusterMarkerIconView(context, densityType)
-                    marker.icon = OverlayImage.fromView(clusterMarkerIconView)
-                    marker.anchor = PointF(0.5F, 0.5F)
-                    marker.captionText = info.size.toString()
-                    marker.captionColor = captionColor.toArgb()
-                    marker.onClickListener = Overlay.OnClickListener {
-                        marker.map?.let { map ->
-                            setCameraToMarker(
-                                map = map,
-                                clickedMarkerPosition = marker.position
-                            )
-                            mapViewModel.setClickedMarkerState(
-                                context = context,
-                                marker = marker,
-                                clusterTag = info.tag.toString()
-                            )
-                        }
-                        true
-                    }
-                    // 클러스터 마커를 클릭한 채로 configuration change 시 크기 유지
-                    if (info.tag.toString()
-                        == mapViewModel.clickedMarkerState.value.clusterPickList?.joinToString(",") { it.id }
-                    ) {
-                        mapViewModel.setClickedMarker(context, marker)
-                    }
-                }
-            })
-            .leafMarkerUpdater(object : DefaultLeafMarkerUpdater() {
-                override fun updateLeafMarker(info: LeafMarkerInfo, marker: Marker) {
-                    marker.anchor = Marker.DEFAULT_ANCHOR
-                    marker.captionText = ""
+        clusterer = buildClusterer(context, mapViewModel)
 
-                    val pick = (info.key as MarkerKey).pick
-                    val leafMarkerIconView = LeafMarkerIconView(context).apply {
-                        val color = if (pick.createdBy.userId == mapViewModel.getUserId()) Blue else Primary
-                        setPaintColor(color.toArgb())
-                    }
-                    leafMarkerIconView.setLeafMarkerIcon(pick) {
-                        marker.icon = OverlayImage.fromView(leafMarkerIconView)
-                        marker.setOnClickListener {
-                            marker.map?.let { map ->
-                                setCameraToMarker(
-                                    map = map,
-                                    clickedMarkerPosition = marker.position
-                                )
-                                mapViewModel.setClickedMarkerState(
-                                    context = context,
-                                    marker = marker,
-                                    pickId = pick.id
-                                )
-                            }
-                            true
-                        }
-                        // 단말 마커를 클릭한 채로 configuration change 시 크기 유지
-                        if (pick.id == mapViewModel.clickedMarkerState.value.curPickId) {
-                            mapViewModel.setClickedMarker(context, marker)
-                        }
-                    }
-                }
-            })
-            .build()
         onDispose {
             clusterer?.clear()
         }
@@ -251,7 +138,7 @@ fun NaverMap(
     )
 }
 
-private fun setCameraToMarker(
+internal fun setCameraToMarker(
     map: NaverMap,
     clickedMarkerPosition: LatLng
 ) {
